+++ src/jadeitite/config.h
@@ -6,10 +6,10 @@
 //======================================
 
 #define JADEITITE_VER_MAJOR 0
-#define JADEITITE_VER_MINOR 2
-#define JADEITITE_VER_PATCH 1
+#define JADEITITE_VER_MINOR 3
+#define JADEITITE_VER_PATCH 0
 
 // Framework version date YYYYMMDD
-#define JADEITITE_VER_DATE 20220626
+#define JADEITITE_VER_DATE 20220628
 
 #endif //JADEITITE_CONF_H