+++ src/jadeitite/datatypes.h
@@ -1,6 +1,8 @@
 #ifndef JADEITITE_DT_H
 #define JADEITITE_DT_H
 
+#include "jadeitite/sys.h"
+
 //======================================
 //           Data types
 //======================================
@@ -8,15 +10,21 @@
 typedef unsigned char u8;
 typedef unsigned short u16;
 typedef unsigned int u32;
+#ifdef SYS64
 typedef unsigned long u64;
+#endif
 
 typedef char s8;
 typedef short s16;
 typedef int s32;
+#ifdef SYS64
 typedef long s64;
+#endif
 
 typedef float f32;
+#ifdef SYS64
 typedef double f64;
+#endif
 
 typedef u8 byte;
 typedef u16 word;
@@ -26,14 +34,11 @@ typedef u32 dword;
 //           Enumerators
 //======================================
 
-typedef enum {
-  RS_OK = 0,
-  RS_FAILED = 1,
-
-  RS_MEMORY_OVERFLOW = 100,
-
-  RS_UNKNOWN_STATUS = 999,
-} return_status_e;
+/**
+ * Used for function return flag as u8/byte
+ */
+#define RS_OK     0
+#define RS_FAILED 1
 
 //======================================
 //             Structures
@@ -79,9 +84,11 @@ typedef struct {
   f32 x, y;
 } vec_2_f32;
 
+#ifdef SYS64
 typedef struct {
   f64 x, y;
 } vec_2_f64;
+#endif
 
 //--------------------------------------
 //               Vector 3
@@ -99,6 +106,12 @@ typedef struct {
   u32 x, y, z;
 } vec_3_u32;
 
+#ifdef SYS64
+typedef struct {
+  u64 x, y, z;
+} vec_3_u64;
+#endif
+
 typedef struct {
   s8 x, y, z;
 } vec_3_s8;
@@ -111,12 +124,20 @@ typedef struct {
   s32 x, y, z;
 } vec_3_s32;
 
+#ifdef SYS64
+typedef struct {
+  s64 x, y, z;
+} vec_3_s64;
+#endif
+
 typedef struct {
   f32 x, y, z;
 } vec_3_f32;
 
+#ifdef SYS64
 typedef struct {
   f64 x, y, z;
 } vec_3_f64;
+#endif
 
 #endif //JADEITITE_DT_H